Fixes #31, allowing unpinning via the webui
There are few fixes in this, mostly though this is to allow users to
unpin from the webui. This commit includes a version bump of the
`node-ipfs-api`, as well as passing in the recursive flag.

An unexpected addition to this commit is due to a change in `go-ipfs`
passing back an object of keys instead of an array.