chore: run tests against dev server in dev (#663)
Our e2e test were pointing at the prod build while our unit tests
would test the current src tree. This quickly leads to annoyance.

This PR runs the e2e style tests against the dev server, and spins
up a test server in ci. This simplifes the run-scripts too.

The only trade off is the test will fail if a server isn't running,
but as the dev flow is to run the dev server and leave the tests
running in watch mode, this isn't an issue.

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>