chore: extract e2e test from unit tests
This change lets us run the unit tests without triggering the e2e tests.
This keeps the unit tests fast and light, to encourage developers to
run them in watch mode as they code.

This also adds react-helmet to control the page title and updates
the e2e navigation test to check for the page title value rather
than an h1 element which is not called for in the designs.

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>