feat: IPLD graph for explore page (#672)
Add a graph to the "explore" page to show the links from the current node and their target node type. It uses [cytoscape](http://js.cytoscape.org/) and [dagre layout](https://github.com/dagrejs/dagre).

It also adds a "graph crumb" component, which shows the node containment and path boundaries that have been traversed along the user provided path.

![graph-crumb-explained](https://user-images.githubusercontent.com/58871/41230393-85add2da-6d77-11e8-8188-b70ba3a5f122.png)

This change required upgrading the logic for resolving info from them the path. Previously we passed the `cid+path` to `dag.get` and tried to render whatever value came back. As dag-cbor node values don't provide the same info as dag-pb nodes _(no `links` or `multihash` properties)_ this pushed complexity on to the UI code of working out how to handle the differences. This PR fixes that by normalising data from `dag.get` into a common shape. The `normaliseDagNode` function currently handles `dag-pb` and `dag-cbor` but the plan is it can be extended to normalise _all the things._

`resolveIpldPath` gives the additional info about the path by recursively calling `dag.get` from the root cid to find all the path boundaries (links that have to be traversed) to resolve the entire path. The normalised array of `nodes` and `pathBoundaries` that were crossed are made available to the the UI so the Graph crumb component can visualise it.


License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>