Make linting work
Run standard on the whole project with default args. Running
standard takes a file path glob as it's arg, so `standard src`
wasn't testing anything.

Explicitly disable es-lint on the webpack config. That file comes
from `react-scripts`; we want to keep it as close to the original
as possible so we can manually merge updates from upstream.

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>