Edit a mutable local copy of the node config
We fetch the node config from the api. This is stashed in the redux
tree, and is only updated by re-fetching from the api.

The settings page manges a local mutable copy that is initialised
on first render. The save button becomes active when there is a
change in the editor. If the user reverts their changes the save
button becomes disabled again to indicate in a calm way that there
are no changes.

In the background we periodically refetch the config from the api.
If we detect that the config was changed since the user started
looking at it in the webui, we show a warning and invite them to
reset the editor to see the changes.

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>