chore: reuse ipld-explorer as git submodule
This is the shortest path to re-using the ipld explorer code
in webui. Publishing ipld explorer as a reuseable module needs
more thought, and right now we want to be able to build and
edit all of the webui code from a single code base, while we
do the polishing for release.

In the future ipld explorer may become one app in a learna repo
for all the webui sections, or we might focus on a post-message
api to co-ordinate ipfs apps via iframes with webui orchestrating
things and providing the ipfs instance to child frames.

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>