chore: use ipld-explorer-components (#781)
* chore: use ipld-explorer-components

Reuse the ipld-explorer code.

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>

* chore: remove low-value failing test, bump up max bundle size

We have to bump up the max bundle size as the old explore code
wasn't using `ipld` which is a large dep. Fixing that is a
seperate issue.

The App test is failing as it imports code from ipld-explorer
that fancy things like dynamic import(). The test wasn't
actually checking anything meaningful, and fixing the jest
config is gonna take a while as it's managed by CRA. Hence
this PR marks it as skipped, as a reminder to figure it out
later.

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>

* chore: fix e2e test

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>