feat: add footer with code and bugs links
Add a footer on the status page with links to the repo and bug reporing.

Only add it on the stats page as it looks weird during page load on the
other pages... it needs more design consideration before we roll it out
across all of them.

This isn't using our shared component library yet, but it can in the future

Aldo adds some layout tweaks on satatus page for large screens.

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>