fix: skip window.ipfs
Web UI wants all the privileges. With window.ipfs the files view
gets silo'd to the files added _in this origin_ so it's gonna
cause confusion.

This PR updates thing to use the new ipfs-redux-bundle and skips
window.ipfs in the provider checks.

Also adds an identity bundle as we no longer use /id as the
ipfs connection test, to move us towards supporting the read-only
api where it is all that is available.

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>