fix: normalise paths when adding to ipfs
Dropped files are given an absolute path, but the "root" is the root
of the dropped directory. Files added via the file input have a
relative path. Normalising them fixes the logic around counting the
number of directories that we expect ipfs will create for us, which
fixes the check we do for sporadic truncation of responses.

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>