fix: handle file-drop inside empty dir
if you drop files inside dirs that have zero or fewer items than
fills the vertical height, the drop would land on the app rather
than the file list, which would cause files dropped inside the dir
to be added at the root rather than the dir.

This PR is a quick fix to ensure the drops handled at the app level
add the files to the current dir where we are on the files page.

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>