chore: fix e2e test
https://github.com/ipfs-shipyard/ipfs-webui/pull/1039 fixed a bunch
of language issues, including upgrading a hyphen to and endash in
the page title of the ipld explore section. The page title is used
to by the e2e test to check that navigation has worked.

This tweak should fix e2e tests for this PR.

We have a more general issue in that we don't get CI tests for PRs
from forks, as circle-ci (and all the CI systems) have no good way
to protect users from exfiltration of secrets if you let anyone
trigger a CI build, and the config for CI is stored in the repo...