feat: prioritize geoip lookup of closest peers
This change improves performance of initial load
of Peers screen when there are thousands of peers.

During the first three times PeerLocationResolver.findLocations is called
we sort peers by latency and resolve geoip only for the closest subset.

This ensures the top of the list in the UI gets updated fast while
thousands of more distant peers gets resolved later.

closes #1273