refactor: reuse know remote pin state
This adds cache for not-pins per service, so we can cache remote pin state
after initial check and avoid sending unnecessary request every time we
enter the same directory.

TODO: on the initial load, Files screen does not run the check for MFS
root. One needs to go to a subdirectory to trigger it, and then if one
goes back to MFS root.