fix(ui): replace IPFS API with Kubo RPC
This aims to finish our UX migration from "ipfs daemon" being IPFS,
and making it explicitly clear that there is no "IPFS API", but we use
"Kubo daemon" and "Kubo RPC API".

Many users are not aware that 99% of their interactions with IPFS
depend on Kubo, and if things go wrong, they are not sure what "Kubo
RPC" is.

This PR replaces "IPFS API" with "Kubo RPC" to make it more clear
what components are at play, and what terms to use when looking for
help.